requirements.txt
streamlit>=1.35.0
groq>=0.9.0
openai>=1.30.0
google-generativeai>=0.7.0
sentence-transformers>=3.0.0
PyMuPDF>=1.24.0
requests>=2.31.0
torch>=2.0.0
duckduckgo-search>=4.0.0
